+++ src/main/java/dev/nahum/nahumInventoryStuff/InventoryTools.java
@@ -6,19 +6,19 @@ import net.minecraft.nbt.NbtIo;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.OfflinePlayer;
-import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.Command;
 import org.bukkit.ChatColor;
+import org.bukkit.command.TabExecutor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
 import java.io.File;
-import java.util.Optional;
+import java.util.*;
 
-public class InventoryTools implements CommandExecutor {
+public class InventoryTools implements TabExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
         if (args.length == 0) {
@@ -44,6 +44,10 @@ public class InventoryTools implements CommandExecutor {
 
             switch (option) {
                 case "see":
+                    if(!sender.hasPermission("nahum.inventorytools.see") && !sender.hasPermission("nahum.inventorytools")){
+                        sender.sendMessage(ChatColor.RED + "You do not have permission to use this command!");
+                        return true;
+                    }
                     if (!(sender instanceof Player)) {
                         sender.sendMessage(ChatColor.RED + "You must be a player to use this command!");
                         return true;
@@ -56,6 +60,10 @@ public class InventoryTools implements CommandExecutor {
                     break;
 
                 case "transfer":
+                    if(!sender.hasPermission("nahum.inventorytools.transfer") && !sender.hasPermission("nahum.inventorytools")){
+                        sender.sendMessage(ChatColor.RED + "You do not have permission to use this command!");
+                        return true;
+                    }
                     if (args.length < 3) {
                         sender.sendMessage(ChatColor.RED + "Usage: /inventorytools transfer <recipient> <giver>");
                         return true;
@@ -74,12 +82,20 @@ public class InventoryTools implements CommandExecutor {
                     break;
 
                 case "clear":
+                    if(!sender.hasPermission("nahum.inventorytools.clear") && !sender.hasPermission("nahum.inventorytools")){
+                        sender.sendMessage(ChatColor.RED + "You do not have permission to use this command!");
+                        return true;
+                    }
                     cleanInventory(targetPlayer, sender);
                     GoodLogger.info(sender.getName() + " has cleared " +
                             targetPlayer.getName() + "'s inventory using the command /inventorytools with the argument \"clear\"!");
                     break;
 
                 case "swap":
+                    if(!sender.hasPermission("nahum.inventorytools.swap") && !sender.hasPermission("nahum.inventorytools")){
+                        sender.sendMessage(ChatColor.RED + "You do not have permission to use this command!");
+                        return true;
+                    }
                     if (args.length < 3) {
                         sender.sendMessage(ChatColor.RED + "Usage: /inventorytools transfer <recipient> <giver>");
                         return true;
@@ -105,6 +121,50 @@ public class InventoryTools implements CommandExecutor {
         }
 
         return false;
+    }
+
+    @Override
+    public List<String> onTabComplete(CommandSender sender, Command cmd, String label, String[] args) {
+        List<String> completions = new ArrayList<>();
+        String currentInput = null;
+        if(args.length == 1) {
+            if(sender.hasPermission("nahum.inventorytools")){
+                completions.add("clear");
+                completions.add("swap");
+                completions.add("transfer");
+                completions.add("see");
+                completions.removeIf(option -> !option.toLowerCase().startsWith(args[0].toLowerCase()));
+                return completions;
+            }
+            if(sender.hasPermission("nahum.inventorytools.see")){
+                completions.add("see");
+            }
+            if(sender.hasPermission("nahum.inventorytools.swap")){
+                completions.add("swap");
+            }
+            if(sender.hasPermission("nahum.inventorytools.transfer")){
+                completions.add("transfer");
+            }
+            if(sender.hasPermission("nahum.inventorytools.clear")){
+                completions.add("clear");
+            }
+            currentInput = args[0].toLowerCase();
+        }
+        if(args.length == 2) {
+            List<String> players = new LinkedList<>();
+            for(OfflinePlayer player : OfflinePlayerSync.getAllPlayers()){
+                players.add(player.getName());
+            }
+            Collections.sort(players);
+            players.removeIf(option -> !option.toLowerCase().startsWith(args[1]));
+            return players;
+        }
+        if(currentInput == null) {
+            return new ArrayList<>();
+        }
+        String unmatched = currentInput;
+        completions.removeIf(option -> !option.toLowerCase().startsWith(unmatched));
+        return completions;
     }
 
     public static boolean cleanInventory(OfflinePlayer player, CommandSender sender){